… ClosedXML.Tests/Resource/Examples/AutoFilter/DateTimeGroupAutoFilter.xlsx
… ClosedXML.Tests/Resource/Examples/Comments/AddingComments.xlsx
… ClosedXML.Tests/Resource/Examples/Loading/ChangingBasicTable.xlsx
… ClosedXML.Tests/Resource/Examples/Misc/AddingDataSet.xlsx
… ClosedXML.Tests/Resource/Examples/Misc/AddingDataTableAsWorksheet.xlsx
… ClosedXML.Tests/Resource/Examples/Misc/AdjustToContents.xlsx
… ClosedXML.Tests/Resource/Examples/Misc/AdjustToContentsWithAutoFilter.xlsx
… ClosedXML.Tests/Resource/Examples/Misc/BasicTable.xlsx
… ClosedXML.Tests/Resource/Examples/Misc/CellValues.xlsx
… ClosedXML.Tests/Resource/Examples/Misc/Collections.xlsx
… ClosedXML.Tests/Resource/Examples/Misc/CopyingWorksheets.xlsx
… ClosedXML.Tests/Resource/Examples/Misc/DataTypes.xlsx
… ClosedXML.Tests/Resource/Examples/Misc/Formulas.xlsx
… ClosedXML.Tests/Resource/Examples/Misc/FormulasWithEvaluation.xlsx
… ClosedXML.Tests/Resource/Examples/Misc/Hyperlinks.xlsx
… ClosedXML.Tests/Resource/Examples/Misc/InsertingData.xlsx
… ClosedXML.Tests/Resource/Examples/Misc/LambdaExpressions.xlsx
… ClosedXML.Tests/Resource/Examples/Misc/SheetProtection.xlsx
… ClosedXML.Tests/Resource/Examples/Misc/ShowCase.xlsx
… ClosedXML.Tests/Resource/Examples/PivotTables/PivotTables.xlsx
… ClosedXML.Tests/Resource/Examples/Ranges/AddingRowToTables.xlsx
… ClosedXML.Tests/Resource/Examples/Ranges/CopyingRanges.xlsx
… ClosedXML.Tests/Resource/Examples/Ranges/CurrentRowColumn.xlsx
… ClosedXML.Tests/Resource/Examples/Ranges/NamedRanges.xlsx
… ClosedXML.Tests/Resource/Examples/Ranges/ShiftingRanges.xlsx
… ClosedXML.Tests/Resource/Examples/Ranges/TransposeRanges.xlsx
… ClosedXML.Tests/Resource/Examples/Ranges/TransposeRangesPlus.xlsx
… ClosedXML.Tests/Resource/Examples/Styles/StyleFont.xlsx
… ClosedXML.Tests/Resource/Examples/Styles/StyleIncludeQuotePrefix.xlsx
… ClosedXML.Tests/Resource/Examples/Styles/StyleNumberFormat.xlsx