Refactor and speedup adjust to content. Instead of using `SixLabors.Fonts` `TextMeasurer`, calculate each grapheme separately (i.e. the way Excel does it).

Adjust first gets box for each grapheme and that is the laid out according to alignment. This PR keeps original functionality, but Excel can lay boxes within a band of certain size and we might support more complex layouts in the future:
![image](https://user-images.githubusercontent.com/7634052/227088329-70ae0640-835d-4a14-9100-fef506ae55ae.png)

It is not as fast, as it used to be, but significant improvement.

The sample for adjust to content from #1991 run using this PR:

|Run# | PR     | 0.100.3 | 0.96.0|
|-----|--------|---------|-------|
|1    | 1674 ms| 1203 ms | 168 ms|
|2    |   31 ms|  568 ms |   4 ms|
|3    |   25 ms|  551 ms |   4 ms|
|4    |   26 ms|  540 ms |   4 ms|
|5    |   25 ms|  544 ms |   3 ms|
|6    |   26 ms|  550 ms |   3 ms|
|7    |   46 ms|  556 ms |   4 ms|
|8    |   28 ms|  556 ms |   3 ms|
|9    |   23 ms|  541 ms |   4 ms|
|10   |   23 ms|  535 ms |   3 ms|

Resolves #1991.